% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate Mean}
\usage{
calculate_mean(data)
}
\arguments{
\item{data}{A data frame with simulated data.}
}
\value{
The mean value.
}
\description{
This function calculates the mean of the simulated data.
}
\examples{
data <- simulate_data(100, 0, 1)
calculate_mean(data)
}
