% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{calculate_sd}
\alias{calculate_sd}
\title{Calculate Standard Deviation}
\usage{
calculate_sd(data)
}
\arguments{
\item{data}{A data frame with simulated data.}
}
\value{
The standard deviation.
}
\description{
This function calculates the standard deviation of the simulated data.
}
\examples{
data <- simulate_data(100, 0, 1)
calculate_sd(data)
}
