% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Data}
\usage{
plot_data(data)
}
\arguments{
\item{data}{A data frame with simulated data.}
}
\value{
A ggplot object.
}
\description{
This function plots the simulated data.
}
\examples{
data <- simulate_data(100, 0, 1)
plot_data(data)
}
