% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data}
\usage{
simulate_data(n, mean, sd)
}
\arguments{
\item{n}{The number of data points to simulate.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}
}
\value{
A data frame with simulated data.
}
\description{
This function simulates a simple dataset.
}
\examples{
simulate_data(100, 0, 1)
}
